refactor : header components and add StageBadge for theme indication
