+++ src/screens/splash/index.ts
@@ -1,10 +1,10 @@
+export { StageBadge } from '../../shared/components/StageBadge';
+export { PackageInfo } from './components/package-info';
 export { InfoSection } from './components/sections/info-section';
 export { LogoSection } from './components/sections/logo-section';
 export { ProgressSection } from './components/sections/progress-section';
-export { PackageInfo } from './components/package-info';
-export { StageBadge } from './components/stage-badge';
-export { VersionInfo } from './components/version-info';
 export { SplashScreen } from './components/splash-screen';
+export { VersionInfo } from './components/version-info';
 export { useSplashViewModel } from './hooks/use-splash-view-model';
 export * from './models/splash-models';
 export { useSplashStore } from './store/splash-store';
