+++ src/shared/components/index.ts
@@ -6,11 +6,12 @@ export { Dropdown } from './Dropdown';
 export { ErrorBoundary } from './ErrorBoundary';
 export { ExternalLink } from './ExternalLink';
 export { HapticTab } from './HapticTab';
-export { LanguageSwitcher } from './LanguageSwitcher';
+//export { LanguageSwitcher } from './LanguageSwitcher';
 export { Loading } from './Loading';
 export { default as ParallaxScrollView } from './ParallaxScrollView';
 export { SafeAnimatedView } from './SafeAnimatedView';
 export { SafeAreaScrollView } from './SafeAreaScrollView';
+export { ScaffoldMessage } from './ScaffoldMessage';
 export { SettingsSection } from './SettingsSection';
 export { ThemedText } from './ThemedText';
 export { ThemedView } from './ThemedView';
